moved tests from SearchWriteBuilderITSpec to WriteConfigITSpec
